#3746 fix: Изменил тип данных proteins, fats, carbohydrates на float

+++ back/database/migrations/2023_06_19_161237_create_dishes_table.php
@@ -19,11 +19,11 @@ return new class extends Migration
                 ->comment('Цена');
             $table->integer('calorie')
                 ->comment('Калории');
-            $table->integer('proteins')
+            $table->float('proteins')
                 ->comment('Белки');
-            $table->integer('fats')
+            $table->float('fats')
                 ->comment('Жиры');
-            $table->integer('carbohydrates')
+            $table->float('carbohydrates')
                 ->comment('Углеводы');
             $table->string('composition')
                 ->comment('Состав');
